Added class Stats. Rewrote old comment. Understood and fixed empty method.
