+++ 2312/2/battlers_table.rb
@@ -1,11 +1,11 @@
 require 'terminal-table'
+require './table_filler'
 
 # BattlersTable prints a table of data, you provide to it
 class BattlersTable
   attr_reader :num_top, :battlers
   def initialize(battlers, num_top)
     @battlers = battlers
-    @counter = -1
     @table = []
     @num_top = num_top
   end
@@ -24,25 +24,17 @@ class BattlersTable
     exit
   end
 
-  def create_table; end
+  def create_table
+    table = []
+    @battlers.sort_by { |_name, info| info.stats.rapper_stats[:bad_words_num] }.reverse.to_h.keys.each do |name|
+      table << TableFiller.new(name, battlers[name]).filled_line
+    end
+    table
+  end
 
   def create_table!
     @battlers.sort_by { |_name, info| info.stats.rapper_stats[:bad_words_num] }.reverse.to_h.keys.each do |name|
-      @counter += 1
-      fill_table(name)
+      @table << TableFiller.new(name, battlers[name]).filled_line
     end
-  end
-
-  def fill_table(name)
-    info = battlers[name]
-    @table[@counter] = [name]
-    bad_words_num = info.stats.rapper_stats[:bad_words_num]
-    @table[@counter] << "#{info.overall[:battles_num]} battles" << "#{bad_words_num} bad words"
-    fill_stats_in_table_from(info)
-  end
-
-  def fill_stats_in_table_from(info)
-    stats = info.stats.rapper_stats
-    @table[@counter] << "#{stats[:bad_words_per_battle]} bad/battle" << "#{stats[:words_per_round]} words/round"
   end
 end
